Make GIN test using injection points repeatable
As written, the test would fail when run repeatedly because one of the
injection points attached was not detached.  This would not matter if
the test is rewritten to be concurrently safe, but let's be clean and
it is a good practice.

Oversight in 6a1ea02c491d.

Discussion: https://postgr.es/m/ZfP7IDs9TvrKe49x@paquier.xyz
(cherry picked from commit 997db123c054c597164be77b7bf7da6479471621)