[OSS-ONLY] Fix crash in InitializeParallelDSM due to MyProcPort not set (#359)
Crash in InitializeParallelDSM during VACUUM by a bg worker InitializeParallelDSM was crashing when a cron job was scheduled to vacuum a table, index or index build/reindexing as there are done by maintenance worker. This change was pushed into serval branches.

This is happenning because for maintenance worker MyProcPort is not set which is why we were getting segfault. With this commit, we have updated the condition to check whether MyProcPort is set or not.

Signed-off-by: Shameem Ahmed shmeeh@amazon.com